Create MatchJoinButton component
