+++ services/frontend/src/components/Tournament/MatchCard.tsx
@@ -1,7 +1,8 @@
 import Babact from "babact";
-import { Match, MatchState } from "../../views/TournamentView";
 import { HPosition, VPosition } from "./Stage";
 import Avatar from "../../ui/Avatar";
+import { useAuth } from "../../contexts/useAuth";
+import { Match, MatchState } from "../../hooks/useTournament";
 
 export default function MatchCard({
 		match,
@@ -25,6 +26,8 @@ export default function MatchCard({
 		return null;
 	}
 
+	const { me } = useAuth();
+
 	return (
 		<div
 			className={`match-card-container flex items-center justify-center ${positionH} ${positionV}`}
@@ -36,14 +39,13 @@ export default function MatchCard({
 				{match.teams.map((team, i) =>
 					<div
 						key={i}
-						className={`match-card-team flex items-center justify-between gap-4 w-full ${getWinnerIndex() === i  ? 'winner' : ''}`}
+						className={`match-card-team flex items-center justify-between gap-4 w-full ${getWinnerIndex() === i  ? 'winner' : ''} ${match.state === MatchState.PLAYING && match.playerTeamIndex(me.account_id) === i ? 'playing' : ''}`}
 					>
 						<div className='flex items-center gap-2'>
 							{team.players.map((player, i) =>
 								<Avatar key={i} src={player.profile.avatar} name={player.profile.username} size='xs'/>
 							)}
-							<h3>{team.name ?? team.players[0].profile.username}</h3>
-							{/* {getWinnerIndex() === i && <i className="fa-solid fa-trophy"></i>} */}
+							<h3>{team.name ?? `${team.players[0].profile.username}'s team`}</h3>
 						</div>
 						<p>{match.scores[i]}</p>
 					</div>
