+++ services/frontend/src/components/Tournament/tournament.css
@@ -122,3 +122,32 @@
 .match-card-team:nth-child(even) {
 	border-right: 2px solid var(--team-2-color);
 }
+
+.match-card-team.playing {
+	animation: background-color-change 2s infinite;
+}
+
+@keyframes background-color-change {
+	0% {
+		background-color: var(--bg-3);
+	}
+	50% {
+		background-color: var(--bg-5);
+	}
+	1000% {
+		background-color: var(--bg-3);
+	}
+}
+
+
+/* ==== Join Button ==== */
+
+.match-join-button {
+	position: absolute;
+	top: calc(var(--header-height) + 16px);
+	left: 50%;
+	transform: translateX(-50%);
+	width: fit-content;
+	z-index: 10;
+	font-size: medium;
+}
