+++ services/frontend/src/views/views.css
@@ -16,13 +16,3 @@
 	align-items: stretch;
 	justify-content: center;
 }
-
-.tournament-join-button {
-	position: absolute;
-	top: calc(var(--header-height) + 16px);
-	left: 50%;
-	transform: translateX(-50%);
-	width: fit-content;
-	z-index: 10;
-	font-size: medium;
-}
