+++ services/nginx/conf/default.conf
@@ -2,7 +2,7 @@
 server {
 	listen 8080 ssl;
 
-	server_name test $DOMAIN_NAME;
+	server_name $DOMAIN_NAME;
 
 	ssl_certificate /run/secrets/ssl-crt;
 	ssl_certificate_key /run/secrets/ssl-key;
@@ -16,7 +16,7 @@ server {
 server {
 	listen 7979 ssl;
 
-	server_name test $DOMAIN_NAME;
+	server_name $DOMAIN_NAME;
 
 	ssl_certificate /run/secrets/ssl-crt;
 	ssl_certificate_key /run/secrets/ssl-key;
